app/requirements.txt
Django==4.1.4
djangorestframework==3.14.0
drf-spectacular==0.25.0
Pillow==9.3.0
matplotlib==3.6.0
seaborn==0.12.1
scipy==1.9.3
pandas==1.5.1
requests==2.28.1